requirements.txt
annotated-types==0.6.0
anyio==4.3.0
app==0.0.1
app-config==1.0.1
Authlib==1.3.0
blinker==1.7.0
boto==2.49.0
cachelib==0.12.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
Choco==1.0.5
click==8.1.7
colorama==0.4.6
config==0.5.1
cryptography==42.0.7
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
Flask==3.0.3
Flask-OAuthlib==0.9.6
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
oauth==1.0.1
oauthlib==2.1.0
openai==1.17.0
orjson==3.10.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
PyMySQL==1.1.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-oauth2==1.1.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.1.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a38
starlette==0.37.2
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
